IGNITE-16932 StoredCacheData logic moved to GridLocalConfigManager
