+++ modules/core/src/main/java/org/apache/ignite/internal/pagemem/store/IgnitePageStoreManager.java
@@ -18,7 +18,6 @@
 package org.apache.ignite.internal.pagemem.store;
 
 import java.nio.ByteBuffer;
-import java.util.Map;
 import java.util.function.Predicate;
 import org.apache.ignite.IgniteCheckedException;
 import org.apache.ignite.configuration.CacheConfiguration;
@@ -26,7 +25,6 @@ import org.apache.ignite.internal.pagemem.PageMemory;
 import org.apache.ignite.internal.processors.cache.CacheGroupContext;
 import org.apache.ignite.internal.processors.cache.CacheGroupDescriptor;
 import org.apache.ignite.internal.processors.cache.GridCacheSharedManager;
-import org.apache.ignite.internal.processors.cache.StoredCacheData;
 import org.apache.ignite.internal.processors.cache.persistence.pagemem.PageMetrics;
 import org.apache.ignite.internal.processors.cache.persistence.pagemem.PageReadWriteManager;
 import org.apache.ignite.internal.processors.cluster.IgniteChangeGlobalStateSupport;
@@ -61,10 +59,10 @@ public interface IgnitePageStoreManager extends GridCacheSharedManager, IgniteCh
      * Callback called when a cache is starting.
      *
      * @param grpDesc Cache group descriptor.
-     * @param cacheData Cache data of the cache being started.
+     * @param ccfg Cache configuration.
      * @throws IgniteCheckedException If failed to handle cache start callback.
      */
-    public void initializeForCache(CacheGroupDescriptor grpDesc, StoredCacheData cacheData)
+    public void initializeForCache(CacheGroupDescriptor grpDesc, CacheConfiguration<?, ?> ccfg)
         throws IgniteCheckedException;
 
     /**
@@ -175,27 +173,6 @@ public interface IgnitePageStoreManager extends GridCacheSharedManager, IgniteCh
      * @throws IgniteCheckedException If failed.
      */
     public int pages(int grpId, int partId) throws IgniteCheckedException;
-
-    /**
-     * @return Saved cache configurations.
-     * @throws IgniteCheckedException If failed.
-     */
-    public Map<String, StoredCacheData> readCacheConfigurations() throws IgniteCheckedException;
-
-    /**
-     * @param cacheData Cache configuration.
-     * @param overwrite Whether stored configuration should be overwritten if it exists.
-     * @throws IgniteCheckedException If failed.
-     */
-    public void storeCacheData(StoredCacheData cacheData, boolean overwrite) throws IgniteCheckedException;
-
-    /**
-     * Remove cache configuration data file.
-     *
-     * @param cacheData Cache configuration.
-     * @throws IgniteCheckedException If failed.
-     */
-    public void removeCacheData(StoredCacheData cacheData) throws IgniteCheckedException;
 
     /**
      * @param grpId Cache group ID.
