+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/CachesRegistry.java
@@ -299,7 +299,7 @@ public class CachesRegistry {
             @Override public void run() {
                 try {
                     for (StoredCacheData data : cacheConfigsToPersist)
-                        cctx.cache().saveCacheConfiguration(data, false);
+                        cctx.cache().configManager().saveCacheConfiguration(data, false);
                 }
                 catch (IgniteCheckedException e) {
                     U.error(log, "Error while saving cache configurations on disk", e);
