+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/GridLocalConfigManager.java
@@ -17,8 +17,22 @@
 
 package org.apache.ignite.internal.processors.cache;
 
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.nio.file.DirectoryStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.Deque;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -26,21 +40,35 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
+import java.util.function.BiConsumer;
 import org.apache.ignite.IgniteCheckedException;
 import org.apache.ignite.IgniteLogger;
 import org.apache.ignite.IgniteSystemProperties;
 import org.apache.ignite.configuration.CacheConfiguration;
 import org.apache.ignite.configuration.IgniteConfiguration;
+import org.apache.ignite.failure.FailureContext;
+import org.apache.ignite.failure.FailureType;
 import org.apache.ignite.internal.GridKernalContext;
 import org.apache.ignite.internal.pagemem.store.IgnitePageStoreManager;
+import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager;
+import org.apache.ignite.internal.processors.cache.persistence.filename.PdsFolderSettings;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.T2;
 import org.apache.ignite.internal.util.typedef.internal.CU;
 import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.lang.IgniteUuid;
+import org.apache.ignite.marshaller.Marshaller;
+import org.apache.ignite.marshaller.MarshallerUtils;
 
+import static java.nio.file.Files.newDirectoryStream;
 import static org.apache.ignite.cache.CacheAtomicityMode.TRANSACTIONAL_SNAPSHOT;
 import static org.apache.ignite.internal.processors.cache.GridCacheUtils.isPersistentCache;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.CACHE_DATA_FILENAME;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.CACHE_DIR_PREFIX;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.CACHE_GRP_DIR_PREFIX;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.TMP_SUFFIX;
 
 /**
  * Responsible for restoring local cache configurations (both from static configuration and persistence).
@@ -50,6 +78,9 @@ public class GridLocalConfigManager {
     /** */
     private final boolean startClientCaches =
         IgniteSystemProperties.getBoolean(IgniteSystemProperties.IGNITE_START_CACHES_ON_JOIN, false);
+
+    /** Listeners of configuration changes e.g. overwrite or remove actions. */
+    private final List<BiConsumer<String, File>> lsnrs = new CopyOnWriteArrayList<>();
 
     /** Caches stop sequence. */
     private final Deque<String> stopSeq = new LinkedList<>();
@@ -66,6 +97,15 @@ public class GridLocalConfigManager {
     /** Context. */
     private final GridKernalContext ctx;
 
+    /** Lock which guards configuration changes. */
+    private final ReentrantReadWriteLock chgLock = new ReentrantReadWriteLock();
+
+    /** Absolute directory for file page store. Includes consistent id based folder. */
+    private File storeWorkDir;
+
+    /** Marshaller. */
+    private final Marshaller marshaller;
+
     /**
      * @param cacheProcessor Cache processor.
      * @param kernalCtx Kernal context.
@@ -73,27 +113,219 @@ public class GridLocalConfigManager {
     public GridLocalConfigManager(
         GridCacheProcessor cacheProcessor,
         GridKernalContext kernalCtx
-    ) {
+    ) throws IgniteCheckedException {
         this.cacheProcessor = cacheProcessor;
         ctx = kernalCtx;
         log = ctx.log(getClass());
+        marshaller = MarshallerUtils.jdkMarshaller(ctx.igniteInstanceName());
+
+        //TODO: execute in start method.
+        final PdsFolderSettings folderSettings = ctx.pdsFolderResolver().resolveFolders();
+
+        storeWorkDir = folderSettings.persistentStoreNodePath();
+
+        U.ensureDirectory(storeWorkDir, "page store work directory", log);
+
+    }
+
+    /**
+     * @param ccfgs List of cache configurations to process.
+     * @param ccfgCons Consumer which accepts found configurations files.
+     */
+    public void readConfigurationFiles(
+        List<CacheConfiguration<?, ?>> ccfgs,
+        BiConsumer<CacheConfiguration<?, ?>, File> ccfgCons
+    ) {
+        chgLock.writeLock().lock();
+
+        try {
+            for (CacheConfiguration<?, ?> ccfg : ccfgs) {
+                File cacheDir = cacheWorkDir(ccfg);
+
+                if (!cacheDir.exists())
+                    continue;
+
+                File[] ccfgFiles = cacheDir.listFiles((dir, name) -> name.endsWith(CACHE_DATA_FILENAME));
+
+                if (ccfgFiles == null)
+                    continue;
+
+                for (File ccfgFile : ccfgFiles)
+                    ccfgCons.accept(ccfg, ccfgFile);
+            }
+        }
+        finally {
+            chgLock.writeLock().unlock();
+        }
+    }
+
+    /**
+     * @return Saved cache configurations.
+     * @throws IgniteCheckedException If failed.
+     */
+    public Map<String, StoredCacheData> readCacheConfigurations() throws IgniteCheckedException {
+        if (ctx.clientNode())
+            return Collections.emptyMap();
+
+        File[] files = storeWorkDir.listFiles();
+
+        if (files == null)
+            return Collections.emptyMap();
+
+        Map<String, StoredCacheData> ccfgs = new HashMap<>();
+
+        Arrays.sort(files);
+
+        for (File file : files) {
+            if (file.isDirectory())
+                readCacheConfigurations(file, ccfgs);
+        }
+
+        return ccfgs;
+    }
+
+    /**
+     * @param dir Cache (group) directory.
+     * @param ccfgs Cache configurations.
+     * @throws IgniteCheckedException If failed.
+     */
+    public void readCacheConfigurations(File dir, Map<String, StoredCacheData> ccfgs) throws IgniteCheckedException {
+        if (dir.getName().startsWith(CACHE_DIR_PREFIX)) {
+            File conf = new File(dir, CACHE_DATA_FILENAME);
+
+            if (conf.exists() && conf.length() > 0) {
+                StoredCacheData cacheData = readCacheData(conf);
+
+                String cacheName = cacheData.config().getName();
+
+                if (!ccfgs.containsKey(cacheName))
+                    ccfgs.put(cacheName, cacheData);
+                else {
+                    U.warn(log, "Cache with name=" + cacheName + " is already registered, skipping config file "
+                        + dir.getName());
+                }
+            }
+        }
+        else if (dir.getName().startsWith(CACHE_GRP_DIR_PREFIX))
+            readCacheGroupCaches(dir, ccfgs);
+    }
+
+    /**
+     * @param conf File with stored cache data.
+     * @return Cache data.
+     * @throws IgniteCheckedException If failed.
+     */
+    public StoredCacheData readCacheData(File conf) throws IgniteCheckedException {
+        try (InputStream stream = new BufferedInputStream(new FileInputStream(conf))) {
+            return marshaller.unmarshal(stream, U.resolveClassLoader(ctx.config()));
+        }
+        catch (IgniteCheckedException | IOException e) {
+            throw new IgniteCheckedException("An error occurred during cache configuration loading from file [file=" +
+                conf.getAbsolutePath() + "]", e);
+        }
+    }
+
+    /**
+     * @param conf File to store cache data.
+     * @param cacheData Cache data file.
+     * @throws IgniteCheckedException If failed.
+     */
+    public void writeCacheData(StoredCacheData cacheData, File conf) throws IgniteCheckedException {
+        // Pre-existing file will be truncated upon stream open.
+        try (OutputStream stream = new BufferedOutputStream(new FileOutputStream(conf))) {
+            marshaller.marshal(cacheData, stream);
+        }
+        catch (IOException e) {
+            throw new IgniteCheckedException("An error occurred during cache configuration writing to file [file=" +
+                conf.getAbsolutePath() + "]", e);
+        }
     }
 
     /**
      * Save cache configuration to persistent store if necessary.
      *
-     * @param storedCacheData Stored cache data.
+     * @param cacheData Stored cache data.
      * @param overwrite Overwrite existing.
      */
-    public void saveCacheConfiguration(StoredCacheData storedCacheData, boolean overwrite) throws IgniteCheckedException {
-        assert storedCacheData != null;
+    public void saveCacheConfiguration(
+        StoredCacheData cacheData,
+        boolean overwrite
+    ) throws IgniteCheckedException {
+        assert cacheData != null;
 
         GridCacheSharedContext<Object, Object> sharedContext = cacheProcessor.context();
 
-        if (sharedContext.pageStore() != null
+        boolean shouldStore = sharedContext.pageStore() != null
             && !sharedContext.kernalContext().clientNode()
-            && isPersistentCache(storedCacheData.config(), sharedContext.gridConfig().getDataStorageConfiguration()))
-            sharedContext.pageStore().storeCacheData(storedCacheData, overwrite);
+            && isPersistentCache(cacheData.config(), sharedContext.gridConfig().getDataStorageConfiguration());
+
+        if (!shouldStore)
+            return;
+
+        CacheConfiguration<?, ?> ccfg = cacheData.config();
+        File cacheWorkDir = cacheWorkDir(ccfg);
+
+        cacheProcessor.context().pageStore().checkAndInitCacheWorkDir(ccfg);
+
+        assert cacheWorkDir.exists() : "Work directory does not exist: " + cacheWorkDir;
+
+        File file = cacheConfigurationFile(ccfg);
+        Path filePath = file.toPath();
+
+        chgLock.readLock().lock();
+
+        try {
+            if (overwrite || !Files.exists(filePath) || Files.size(filePath) == 0) {
+                File tmp = new File(file.getParent(), file.getName() + TMP_SUFFIX);
+
+                if (tmp.exists() && !tmp.delete()) {
+                    log.warning("Failed to delete temporary cache config file" +
+                        "(make sure Ignite process has enough rights):" + file.getName());
+                }
+
+                writeCacheData(cacheData, tmp);
+
+                if (Files.exists(filePath) && Files.size(filePath) > 0) {
+                    for (BiConsumer<String, File> lsnr : lsnrs)
+                        lsnr.accept(ccfg.getName(), file);
+                }
+
+                Files.move(tmp.toPath(), file.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            }
+        }
+        catch (IOException ex) {
+            ctx.failure().process(new FailureContext(FailureType.CRITICAL_ERROR, ex));
+
+            throw new IgniteCheckedException("Failed to persist cache configuration: " + ccfg.getName(), ex);
+        }
+        finally {
+            chgLock.readLock().unlock();
+        }
+    }
+
+    /**
+     * Remove cache configuration from persistent store.
+     *
+     * @param cacheData Stored cache data.
+     */
+    public void removeCacheData(StoredCacheData cacheData) throws IgniteCheckedException {
+        chgLock.readLock().lock();
+
+        try {
+            CacheConfiguration<?, ?> ccfg = cacheData.config();
+            File file = cacheConfigurationFile(ccfg);
+
+            if (file.exists()) {
+                for (BiConsumer<String, File> lsnr : lsnrs)
+                    lsnr.accept(ccfg.getName(), file);
+
+                if (!file.delete())
+                    throw new IgniteCheckedException("Failed to delete cache configuration: " + ccfg.getName());
+            }
+        }
+        finally {
+            chgLock.readLock().unlock();
+        }
     }
 
     /**
@@ -136,6 +368,103 @@ public class GridLocalConfigManager {
     }
 
     /**
+     * @param lsnr Instance of listener to add.
+     */
+    public void addConfigurationChangeListener(BiConsumer<String, File> lsnr) {
+        assert chgLock.isWriteLockedByCurrentThread();
+
+        lsnrs.add(lsnr);
+    }
+
+    /**
+     * @param lsnr Instance of listener to remove.
+     */
+    public void removeConfigurationChangeListener(BiConsumer<String, File> lsnr) {
+        lsnrs.remove(lsnr);
+    }
+
+    /**
+     * Delete caches' configuration data files of cache group.
+     *
+     * @param ctx Cache group context.
+     * @throws IgniteCheckedException If fails.
+     */
+    public void removeCacheGroupConfigurationData(CacheGroupContext ctx) throws IgniteCheckedException {
+        File cacheGrpDir = cacheWorkDir(ctx.sharedGroup(), ctx.cacheOrGroupName());
+
+        if (cacheGrpDir != null && cacheGrpDir.exists()) {
+            DirectoryStream.Filter<Path> cacheCfgFileFilter = new DirectoryStream.Filter<Path>() {
+                @Override public boolean accept(Path path) {
+                    return Files.isRegularFile(path) && path.getFileName().toString().endsWith(CACHE_DATA_FILENAME);
+                }
+            };
+
+            try (DirectoryStream<Path> dirStream = newDirectoryStream(cacheGrpDir.toPath(), cacheCfgFileFilter)) {
+                for (Path path: dirStream)
+                    Files.deleteIfExists(path);
+            }
+            catch (IOException e) {
+                throw new IgniteCheckedException("Failed to delete cache configurations of group: " + ctx.toString(), e);
+            }
+        }
+    }
+
+    /**
+     * @param grpDir Group directory.
+     * @param ccfgs Cache configurations.
+     * @throws IgniteCheckedException If failed.
+     */
+    private void readCacheGroupCaches(File grpDir, Map<String, StoredCacheData> ccfgs) throws IgniteCheckedException {
+        File[] files = grpDir.listFiles();
+
+        if (files == null)
+            return;
+
+        for (File file : files) {
+            if (!file.isDirectory() && file.getName().endsWith(CACHE_DATA_FILENAME) && file.length() > 0) {
+                StoredCacheData cacheData = readCacheData(file);
+
+                String cacheName = cacheData.config().getName();
+
+                if (!ccfgs.containsKey(cacheName))
+                    ccfgs.put(cacheName, cacheData);
+                else {
+                    U.warn(log, "Cache with name=" + cacheName + " is already registered, skipping config file "
+                        + file.getName() + " in group directory " + grpDir.getName());
+                }
+            }
+        }
+    }
+
+    /**
+     * @param ccfg Cache configuration.
+     * @return Cache configuration file with respect to {@link CacheConfiguration#getGroupName} value.
+     */
+    private File cacheConfigurationFile(CacheConfiguration<?, ?> ccfg) {
+        File cacheWorkDir = cacheWorkDir(ccfg);
+
+        return ccfg.getGroupName() == null ? new File(cacheWorkDir, CACHE_DATA_FILENAME) :
+            new File(cacheWorkDir, ccfg.getName() + CACHE_DATA_FILENAME);
+    }
+
+    /**
+     * @param ccfg Cache configuration.
+     * @return Store dir for given cache.
+     */
+    private File cacheWorkDir(CacheConfiguration<?, ?> ccfg) {
+        return FilePageStoreManager.cacheWorkDir(storeWorkDir, FilePageStoreManager.cacheDirName(ccfg));
+    }
+
+    /**
+     * @param isSharedGroup {@code True} if cache is sharing the same `underlying` cache.
+     * @param cacheOrGroupName Cache name.
+     * @return Store directory for given cache.
+     */
+    public File cacheWorkDir(boolean isSharedGroup, String cacheOrGroupName) {
+        return FilePageStoreManager.cacheWorkDir(storeWorkDir, FilePageStoreManager.cacheDirName(isSharedGroup, cacheOrGroupName));
+    }
+
+    /**
      * @return {@code True} if need locally start all existing caches on client node start.
      */
     private boolean startAllCachesOnClientStart() {
@@ -168,7 +497,7 @@ public class GridLocalConfigManager {
         }
 
         if (CU.isPersistenceEnabled(config) && pageStoreManager != null) {
-            Map<String, StoredCacheData> storedCaches = pageStoreManager.readCacheConfigurations();
+            Map<String, StoredCacheData> storedCaches = readCacheConfigurations();
 
             if (!F.isEmpty(storedCaches)) {
                 List<String> skippedConfigs = new ArrayList<>();
