+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/GridCacheDatabaseSharedManager.java
@@ -1624,7 +1624,11 @@ public class GridCacheDatabaseSharedManager extends IgniteCacheDatabaseSharedMan
                 CacheGroupContext grp = tup.get1();
 
                 try {
-                    cctx.pageStore().shutdownForCacheGroup(grp, tup.get2());
+                    boolean destroy = tup.get2();
+                    cctx.pageStore().shutdownForCacheGroup(grp, destroy);
+
+                    if (destroy)
+                        cctx.cache().configManager().removeCacheGroupConfigurationData(grp);
                 }
                 catch (IgniteCheckedException e) {
                     U.error(log, "Failed to gracefully clean page store resources for destroyed cache " +
